Reworked the ERB benchmark to not eval.
While ERB requires eval to truly run, the eval operation confounds the
results for String performance. Instead, we now capture the code fragment
that ERB would eval and run it in the benchmark harness.

Since this operation is both reduced is scope and generally quite fast,
I moved to using the microbenchmark harness so we can tweak the iteration
count.